requirements.txt
django
djangorestframework
django-cors-headers
Pillow
joblib
requests
numpy
librosa
ffmpeg-python
pydub
django-extensions
channels
daphne